+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add documentation build with sphinx, myst-nb, and the spinx-book-theme. Also hosted on [read the docs](https://aimmd.readthedocs.io/en/latest/).
+
 ## [0.9.2] - 2025-07-29
 
 ### Added